Revert "Fix for code-preview"
This reverts commit 620501e02cc337faa7171c5a6fa1398eb0db7282.
