--- libs/documentation/src/lib/components/button-group/demos/button-group-modes/button-group-modes.component.ts
@@ -1,7 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  templateUrl: './button-group-modes.component.html',
-  selector: `sds-button-group-modes`,
-})
-export class ButtonGroupModesComponent {}
