--- libs/documentation/src/lib/components/button-group/demos/button-group-modes/button-group-modes.module.ts
@@ -1,12 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-import { SdsButtonGroupModule } from '@gsa-sam/sam-material-extensions';
-import { ButtonGroupModesComponent } from './button-group-modes.component';
-
-@NgModule({
-  imports: [CommonModule, SdsButtonGroupModule],
-  declarations: [ButtonGroupModesComponent],
-  exports: [ButtonGroupModesComponent],
-  bootstrap: [ButtonGroupModesComponent],
-})
-export class ButtonGroupModesModule {}
