+++ libs/documentation/src/lib/components/button-group/demos/modes/button-group-modes.component.html
@@ -1,0 +1,23 @@
+<sds-button-group [mode]="'radio'" class="sds-button-group">
+  <sds-button-group-option value="reports" [aria-label]="'reports'">
+    Reports
+  </sds-button-group-option>
+  <sds-button-group-option value="Radio Mode" [aria-label]="'Radio Mode'">
+    Radio Mode
+  </sds-button-group-option>
+  <sds-button-group-option value="history" [aria-label]="'history'">
+    History
+  </sds-button-group-option>
+</sds-button-group>
+
+<sds-button-group [mode]="'checkbox'" class="sds-button-group">
+  <sds-button-group-option value="reports" [aria-label]="'reports'">
+    Reports
+  </sds-button-group-option>
+  <sds-button-group-option value="Checkbox Mode" [aria-label]="'Checkbox Mode'">
+    Checkbox Mode
+  </sds-button-group-option>
+  <sds-button-group-option value="history" [aria-label]="'history'">
+    History
+  </sds-button-group-option>
+</sds-button-group>
