+++ package.json
@@ -89,8 +89,6 @@
     "@angular/language-service": "^12.2.15",
     "@babel/core": "^7.17.10",
     "@babel/preset-env": "^7.4.5",
-    "@etchteam/storybook-addon-css-variables-theme": "^1.2.1",
-    "@storybook/addon-a11y": "^6.4.22",
     "@storybook/addon-actions": "^6.4.22",
     "@storybook/addon-essentials": "^6.4.22",
     "@storybook/addon-interactions": "^6.4.22",
@@ -105,16 +103,16 @@
     "@types/node": "^14.0.4",
     "babel-loader": "^8.2.5",
     "codelyzer": "^6.0.0",
-    "css-loader": "^6.5.1",
+    "css-loader": "^6.2.0",
     "dotenv": "6.2.0",
-    "ejs": "^3.1.6",
+    "ejs": "3.1.5",
     "eslint": "6.8.0",
     "github-release-notes": "^0.17.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
     "karma": "~6.3.4",
     "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "^1.5.0",
     "ng-packagr": "^12.2.6",
@@ -124,7 +122,7 @@
     "react": "^16.4.2",
     "react-dom": "^16.4.2",
     "storybook-addon-preview": "^2.2.0",
-    "ts-node": "~8.3.0",
+    "ts-node": "~7.0.0",
     "tslint": "~6.1.3"
   }
 }
